feat: :sparkles: abstractions made to main bot file
